style: 컬러 정의 및 파일 이름 변경(#18)
