+++ SOPKATHON-iOS-TEAM2/SOPKATHON-iOS-TEAM2/Common/Resources/Colors/DoYouKnowThisColors.swift
@@ -1,0 +1,44 @@
+//
+//  MockColors.swift
+//  SOPKATHON-iOS-TEAM2
+//
+//  Created by 한지석 on 5/16/24.
+//
+
+import UIKit
+
+extension UIColor {
+    static var gray05: UIColor {
+        return UIColor(white: 242.0 / 255.0, alpha: 1.0)
+    }
+    static var gray06: UIColor {
+        return UIColor(white: 246.0 / 255.0, alpha: 1.0)
+    }
+    static var gray04: UIColor {
+        return UIColor(white: 227.0 / 255.0, alpha: 1.0)
+    }
+    static var gray02: UIColor {
+        return UIColor(white: 153.0 / 255.0, alpha: 1.0)
+    }
+    static var gray01: UIColor {
+        return UIColor(white: 102.0 / 255.0, alpha: 1.0)
+    }
+    static var black: UIColor {
+        return UIColor(white: 34.0 / 255.0, alpha: 1.0)
+    }
+    static var white: UIColor {
+        return UIColor(white: 1.0, alpha: 1.0)
+    }
+    static var p500: UIColor {
+        return UIColor(hexCode: "8000ff")
+    }
+    static var p400: UIColor {
+        return UIColor(hexCode: "a245ff")
+    }
+    static var p300: UIColor {
+        return UIColor(hexCode: "c183ff")
+    }
+    static var p200: UIColor {
+        return UIColor(hexCode: "dfbfff")
+    }
+}
