--- SOPKATHON-iOS-TEAM2/SOPKATHON-iOS-TEAM2/Common/Resources/Colors/MockColors.swift
@@ -1,32 +1,0 @@
-//
-//  MockColors.swift
-//  SOPKATHON-iOS-TEAM2
-//
-//  Created by 한지석 on 5/16/24.
-//
-
-import UIKit
-
-extension UIColor {
-    static var gray05: UIColor {
-      return UIColor(white: 242.0 / 255.0, alpha: 1.0)
-    }
-    static var gray06: UIColor {
-      return UIColor(white: 246.0 / 255.0, alpha: 1.0)
-    }
-    static var gray04: UIColor {
-      return UIColor(white: 227.0 / 255.0, alpha: 1.0)
-    }
-    static var gray02: UIColor {
-      return UIColor(white: 153.0 / 255.0, alpha: 1.0)
-    }
-    static var gray01: UIColor {
-      return UIColor(white: 102.0 / 255.0, alpha: 1.0)
-    }
-    static var black: UIColor {
-      return UIColor(white: 34.0 / 255.0, alpha: 1.0)
-    }
-    static var white: UIColor {
-      return UIColor(white: 1.0, alpha: 1.0)
-    }
-}
